# openapi-station-service.yaml
openapi: 3.0.3
info:
  title: StationService API
  version: 1.0.0
  description: API for ingesting and querying EV charging stations
servers:
  - url: http://localhost:8080
paths:
  /api/v1/stations/upload:
    post:
      summary: Upload station inventory CSV file
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Processing summary (synchronous)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingSummary'
        '202':
          description: Accepted for async processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingSummary'
        '400':
          description: Bad request (malformed file)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/stations/batch:
    post:
      summary: Upload small station batch as JSON array
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/StationDTO'
      responses:
        '200':
          description: Processing summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingSummary'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/stations/{stationId}:
    get:
      summary: Get station by id
      parameters:
        - name: stationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Station internal model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    StationDTO:
      type: object
      properties:
        stationId:
          type: string
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        capacity:
          type: integer
        status:
          type: string
          enum: [ACTIVE, MAINTENANCE, OFFLINE]
      required: [stationId, name, capacity]
    Station:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        displayName:
          type: string
        location:
          type: object
          properties:
            lat:
              type: number
            lon:
              type: number
        capacity:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    ProcessingSummary:
      type: object
      properties:
        processingId:
          type: string
        receivedCount:
          type: integer
        successCount:
          type: integer
        failedCount:
          type: integer
        sampleErrors:
          type: array
          items:
            $ref: '#/components/schemas/RecordError'
    RecordError:
      type: object
      properties:
        recordIndex:
          type: integer
        errorCode:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        processingId:
          type: string
        failedRecords:
          type: integer
        sampleErrors:
          type: array
          items:
            $ref: '#/components/schemas/RecordError'